requirements.txt
Package                       Version
----------------------------- --------
aiohttp                       3.8.4
aiohttp-retry                 2.8.3
aiosignal                     1.3.1
asgiref                       3.6.0
async-timeout                 4.0.2
attrs                         23.1.0
backports.zoneinfo            0.2.1
certifi                       2023.5.7
charset-normalizer            3.1.0
distlib                       0.3.6
Django                        4.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
filelock                      3.12.0
frozenlist                    1.3.3
idna                          3.4
multidict                     6.0.4
phonenumbers                  8.13.11
Pillow                        9.5.0
pip                           23.1.2
platformdirs                  3.5.1
psycopg2                      2.9.6
PyJWT                         2.7.0
python-decouple               3.8
pytz                          2023.3
requests                      2.30.0
setuptools                    47.1.0
sqlparse                      0.4.4
twilio                        8.2.0
typing_extensions             4.6.2
tzdata                        2023.3
urllib3                       2.0.2
virtualenv                    20.23.0
yarl                          1.9.2
